Change Screen: Allow to configure display of size details in file list
Rename the "Show Change Sizes As Colored Bars In Changes Table" setting
to "Show Change Sizes As Colored Bars" and use it to also control how
the change size is shown per file in the file table.

When enabled (which is the default), the change size per file is shown
as a sum of lines added/removed, and also representated by a colored
bar showing the proportion of added/removed lines.

When disabled, the colored bar is not shown and the change size per file
is shown in the same way as it used to be in the old change screen:

- For modified and renamed files, the added and removed line counts are
  shown separately, e.g. "+10, -2".

- For added files, the number of lines in the new file is shown, e.g.
  "100 line".

- No size is shown for deleted files.

Bug: Issue 3148
Change-Id: Iac383edda9ab0c683846046eeedf008fdc606a71

Former-commit-id: 8bd6f7856e6a999ae46978885158af2f7296a48b