Remove documentation of "Edit Message" feature
The "Edit Message" button was removed from the Change Screen; the
commit message is now edited in the inline editor.

Remove the documentation of "Edit Message" from the Review UI
documentation.

Note that some screen shots still show the "Edit Message" button
and these will need to be updated too.

Change-Id: I35f74b9c28cacda39e801269d39d42be5ba94030

Former-commit-id: 964f6b8ff6e0336a835788012be2ec9445828e6a