Only create new changes as draft if draft changes are allowed
In change I33a3cccb the behavior of the "Create Change" and "Follow
Up" buttons was changed so that new changes are always created as
drafts.

The `change.allowDrafts` setting was not considered, which allowed a
new draft change to be created even when disabled. It was then not
possible to publish the draft change because the publish draft action
checks the `change.allowDrafts` setting.

Change it so that changes are only created as draft when draft changes
are allowed. When not allowed, create regular new changes.

Change-Id: Ie8bea86218d23d2efb30afee4500e4eb8c59e2c2

Former-commit-id: 0a1da0981ba40314250946146162b6f6b23685cb