ChangeScreen: Allow to disable muting of common paths in file table
In the file table, parts of the file path that are common to the file
previously listed are muted.

The purpose of this is to make it easier to see files that all belong
under the same path, but some users find it annoying.

Add a new user preference to control this feature. When it is disabled,
the entire path is shown in the same color for all files.

The default setting for the new preference is enabled, meaning the
default behavior is the same as now, and the user must explicitly
disable it.

Bug: Issue 3147
Change-Id: Id9bf7270e3eb65c22c4d44f5ef6717c952aeeb2a

Former-commit-id: 751441fdfaf1b7b9e494118018bc1b798bd29118