Allow to publish draft patch sets even when `allowDrafts` is false
It is possible that a user uploads a draft change while `allowDrafts`
is enabled, but then the administrator disables it.

In this case, the user cannot publish the change and it is stuck in
the draft state.

Remove this restriction so that the user is able to publish the 'stuck'
change.

The similar restriction on deleting drafts is not removed by this
change. Deleting a draft is a destructive operation, so it is reasonable
to keep this restriction in place.

Change-Id: Id38962f933726f5523e0d6d9f318f30d162a910a

Former-commit-id: df99e12530502d9a01f845e21259fa76e93a4e25