Fix: REST API allows to create draft change when drafts are disabled
It was possible to create a draft change via the REST API even when
the `change.allowDrafts` setting was false in the configuration.

Add a check to prevent this.

Extend the acceptance test case to include a configuration with the
`change.allowDrafts` set to false, and add a new test case to verify
that draft changes cannot be created.

Change-Id: I79cff3cd05361e2e81ba2ab51381c9a8c0f04530

Former-commit-id: 6f8beaa9735539f5a589d2042ffbc23df404724a