Fix: REST API projects endpoints do not correctly handle '.git' suffix
The projects REST API documentation states that the `.git` suffix will
be stripped off the input project name, if present.

This was working for the "Create Project" endpoint, because the suffix
is stripped off by the repository manager. However it was not being
stripped off for other endpoints.  This caused the projects API to fail
when called with a project name with `.git` suffix.

Fix the name parsing in the projects collection to strip off the suffix.

Add a new acceptance test class for the "Get Project" endpoint, testing
the cases with and without the `.git` suffix, and with a non-existing
project.

Extend the "Create Project" REST API and projects API acceptance tests
to include the case with `.git` suffix.

Bug: Issue 3153
Change-Id: Id9c851972017e8ff833686c40623c3c4228cad30

Former-commit-id: 3695e7ebc9ad7bc3c0bf495e26ee07acb266f132