RFC: Enable Eclipse warning about fall-through cases in switch statements
Enable the "'switch' case fall-through" warning in Eclipse to catch
unintentional fall-through to the following 'case' block when a 'break'
statement is omitted.

All warnings raised in the current codebase are caused by intentional
fall-throughs, so add "//$FALL-THROUGH$" comments to suppress those
warnings.

This change was inspired by the coverity report [1] which found the
problem in PutConfig (see CID 19880) but interestingly did not report
anything for the other ones.

RFC because I'm not sure if the "//$FALL-THROUGH$" comments will cause
the warnings to be suppressed on subsequent coverity scans.

[1] https://scan.coverity.com/projects/4032

Change-Id: Iad003bd7d086c67cccdbfb1312c8bd82d17fcc57

Former-commit-id: 9bbffa71b3674e0f39db147e14db76a7db9fe474