Fix documentation of "Submit" permission
The documentation mentions the "Submit Patch Set n" button.  It was
named like this in the old change screen, but in the new change screen
it is simply "Submit".

In the current version of Gerrit it is also possible to submit a change
via means other than the UI, so rather than fixing the name of the
button just reword the sentence to simply say that the permission allows
the user to submit changes.

Change-Id: Ie0671a163c875f66744bd2cb731f673ce62af76f

Former-commit-id: 6bf46eddab741fd645575ce47db1f3f3e2389b5a