InlineEdit: Fix diff against revision feature
Since Iccbe4868 diff against revision for both review and edit mode was
broken. Call of REST endpoint to retrieve file list was erroneously
skipped for change edits. It was done to skip drafts and comments
retrieval. Which is safe thing to skip because comments cannot be added
and published for change edits.

Note, that diff against revsion feature is preserved in both modes:
review mode and edit mode. It was discussed in this thread: [1].

[1] https://groups.google.com/d/topic/repo-discuss/20ge_59v3WI/discussion

Bug: Issue 3142
Change-Id: I567a51ed5dca7774ef886b4c9953b14cafa84c46

Former-commit-id: d329a57745cf9c5f00649aedc391dd8a9d931eb0