Throw IOException from JarScanner constructor
The constructor of JarScanner is throwing an InvalidPluginException
when an IOException occurs while loading a JAR file. This exception is
too specific here and it should be the responsibility of the caller to
decide whether a given JAR file is a plugin or not. Therefore this
constructor should throw a generic IOException instead.

Change-Id: I60099f9cec0837e6bf9a72af0d085f9731495c14
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: 5436ff9c233be1c0ef0216185fe473a09c858f92