InlineEdit: Fix test for mangling of non-ASCII characters
FindBugs detected that the test was passing a byte array into a string
concatenation, which causes toString being called on the array, which
returns a string such as `[C@16f0472` rather than the actual contents.

The solution suggested by FindBugs is to use `Arrays.toString` instead,
but this also does not seem to result in the actual content being
appended.

Replace the byte array with a String.

Change-Id: Iaea83a52b405a9fcc615ac8e5a7d6ca93a123077

Former-commit-id: 30c7dac28b5eedafe889770036770fc5676ab7a1