Submit whole topic: disable submit button if not ready
The current implementation just disables the submit button if the current
user could not submit all changes in the topic. Additionally we want to
block the submit button if the changes do not pass the submit rule checks
individually. When the submission is disabled, show a tooltip explaining
why there is a problem.

This change is also still not what we want to have eventually:
We're having a cache invalidation problem, when other changes in
the same topic change and affect the submission of the whole topic here.

Change-Id: Ic242e738efa9b53cbf0dfd167e64b070c84dac29

Former-commit-id: 252d769b51d5248ebedb32f6773dd3dd30e2bf62