Add new REST API call for querying revision actions
If you have enabled the new option `submitWholeTopic`, we eventually want
to query all changes in the topic if they can be submitted (Code-Review +2,
Verified and the user has permission to submit) to determine if the
submit button is enabled in a ChangeScreen. This query may take some
time, so we don't want to have it in the same query for the whole
change screen, which should render fast.

This commit introduces a new REST API call which allows to query
for the available actions for a given revision id:

  changes/<id>/revisions/<id>/actions

By introducing a new ActionJson class we can move all actions related
JSON processing out of ChangeJson eventually which currently
covers actions such as in the change/<id>/detail REST API call.

Change-Id: I58c67944e756f0b130a9c1d89d83f7635bf474a7

Former-commit-id: c725966f8b65220879cede7ad2b45d09311bc951