Add documentation about plugins
A new documentation page explains how to install plugins and how a
project for plugin development can be requested for gerrit-review.
In addition it lists all known plugins with a brief description and
links to the project and the plugin documentation. The plugin
descriptions are already available in each plugin, but it's much
easier to discover plugins, if all descriptions are available on a
single page.

Change-Id: I596423cacf623039367eeaf8f5833ce0e1a831ed
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 6a78963c0cef66a22ce2ee9452f5b72d3bb3d509