Revert "Perform user.getEffectiveGroups() less eagerly"
This reverts commit 78c2d6bda6e63bede53d6e8730d3f47e56c6380c [formerly f7569d0cb27df67728b4c717ba5e4ac2bb4b7080].
The implementation removes any ability for slow GroupBackends
to perform their own per-request group level caching.

The correct way to do less eager lookups is for the LDAP
implementation of GroupBackend to create an emtpy membership
and then populate it on demand as requests arrive and need
to be looked up in the contains() methods.

Change-Id: I65210b24bffe10c5b158668867c8167c84712d17

Former-commit-id: 7ab135e8630792e4cba2902d7a263127305619e7