CherryPick: Keep going in the face of a single merge conflict
In addition to verifying this behavior in a test, this also fixes an
issue caused by setting mergeTip to null. On the next iteration of the
loop we were incorrectly treating the null merge tip as a cherry-pick
onto an unborn root. Setting the commit as CLEAN_MERGE resulted in
rejecting the later ref update in MergeOp.

Change-Id: I82299ea96725a21430a0871199cde3e2f3af267b

Former-commit-id: 8cbf929a206894fe6965629b05bf436a4cd41f22