ChangeScreen: Avoid race condition for revision changes
loadConfigInfo called renderChangeInfo which called renderRevisionInfo with
an empty argument list to initialise the the revision action buttons.
The call to renderChangeInfo however was done via a callback, such that
we are not sure how long it takes until renderChangeInfo is called.

By moving loadRevisionInfo which load the actual revision action info
into that callback we can be sure to not have a race of the callback
for the callchain as described above and the callback in loadRevisionInfo
to render the correct revision info.

This will add one round trip time, however it actually fixes a bug.

Bug: issue 3192
Change-Id: I186ca5306a29e785981c809c92d51f7a0fb3e71a

Former-commit-id: 3bb4767be7151ce75e40e18b52a59352e4821c37