Revert changes from topic 'submitWholeTopic'
Recently merged changes related to the 'submitWholeTopic' feature have
introduced regressions in the change screen.

This feature is not targetted for the 2.11 release. Instead of waiting
for fixes on master, just revert the changes on the stable-2.11 branch.

* This reverts the following commits:
  dbefbe0: Merge changes from topic 'submitWholeTopic'
  3bb4767: ChangeScreen: Avoid race condition for revision changes
  6287784: Submit Whole Topic: trigger merge queue for all submitted changes
  b9db274: ChangeScreen: explicitly load revision info
  46c319e: Actions: reloadRevisionActions determines submit button visibility
  a6fa971: ChangeScreen: introduce renderRevisionInfo
  3b0b9c3: ChangeAPI: add call to new actions REST API call
  add7038: ETag computation needs to honor changes in other changes

Note that the whole 'submitWholeTopic' series is not reverted; only the
recent changes that caused regression. The remaining parts of the series
are working as expected and only when the setting is explicitly enabled.

Also remove the documentation of the 'submitWholeTopic' setting from the
configuration documentation.

Change-Id: I4b7538bec1820743c4a4ac0c4100f56eb02d5ef2

Former-commit-id: dab88b1ab1ba81924d43766b70c049225aeb4eae