ChangeScreen: reword submit button for single change topics
When `change.submitWholeTopic` is enabled and a topic is set for a change,
we always rendered the submit button with "Submit whole topic". This is
confusing when having only one change in the topic, which can easily happen
when a user wants to revert just one change which was merged as part of a
larger topic before. Then the revert commit keeps the topic, but it's just
one change in the topic.

To avoid user confusion when having just one change in the topic, do not
display it as a submission of the whole topic, but rather treat it as a
change without topic.

The change for one change in a topic has been adapted to reflect the new
behavior. This does not worsen the test situation as there are still
	revisionActionsTwoChangeChangesInTopic
	revisionActionsTwoChangeChangesInTopicReady
which test the point of having the whole topic enabled.

Change-Id: Ic97400a05749d9db4a03eb512a52c84f43403a83

Former-commit-id: 619af2aaca14e87c069c5b2e8bfd9d7946cf40e5