MergeabilityCacheImpl: Only get needed refs
In Loader.load the refs variable is only passed to the method
alreadyAccepted which is in turn only interested in refs/heads/* and
refs/tags/*. So instead of loading all references only load the needed
references in the first place.

This slighty improves performance, especially when a repository has many
refs/changes.

Change-Id: I11f2efb59eb00ef65a7323f9b673e11591da1959

Former-commit-id: 0ea9c35ba003eb32f55f405cb078b7b6dcbff170