ChangeScreen: quick approve button is visible again
When I was looking at the code for the quick approve button

    if (current) {
      quickApprove.set(info, revision, replyAction);
    } else {
      quickApprove.setVisible(false);
    }

I assumed that the quickApprove.set will set the visibility again to true,
as I assumed this code path may be called multiple times. This assumption
was wrong, but it led to the change (ChangeScreen: do not show history
twice, I649af53ea56e63febbb9144a6eae306e2db4269e), where the quick approve
button was made invisible unconditionally on page load, and the assumption
was to have it set visible again by quickApprove.set(...).

This change will explicitly set the visibility of the quick approve button
in any code path of quickApprove.set(...).

Bug: issue 3203
Change-Id: I45f7930c8309eacfd46c1a6db391b946afcc39a8

Former-commit-id: ec9909bc563ee584a94ced77d7a385970373cd6c