Inline Edit: Disable 'Close' button until save is completed
On servers where the save operation can take some time, it's possible
to click on 'Save', and then on 'Close' before the save has finished,
which causes a warning about unsaved changes to pop up.

To prevent this, disable the 'Close' button when 'Save' is pressed, and
then re-enable it when the save operation has finished.

Change-Id: I4e707f6aa341e038f17622556f256475af282519

Former-commit-id: 73c0395fe429a4790cf40ca583d6e63f6a8c030b