Do not return 403 when clicking on Gitweb breadcrumb
A message "Forbidden" was displayed when clicking on any part of the
breadcrumb trail displayed at the top of Gitweb pages. That happened for
projects with parent folders because browsing parent folders is not
supported by Gerrit.

Now when the user clicks on the parent folder, redirect to Gerrit projects
list screen with the parent folder path as the filter.

Change-Id: I86dfb3f29d8da6ee02efc95470673fe70a1f2d3e

Former-commit-id: 53793aece3a387eed54c51cd6cc29520636a426b