Fix broken SubmitByCherryPickIT
Change Ide711cd5 added a test that verifies the review message added
when a change is merged, but this fails for the cherry pick test
because the message is different when the submit strategy is cherry
pick.

Update the test to check for the correct message when the submit
strategy is cherry pick.

Change-Id: If5966e5a5f21c68e52f87e49a842f55f7b578c8c

Former-commit-id: 8541643db5651c5923cf99d67ef178e0d1449d54