Add support to consume CodeMirror from org.webjars on Maven Central
CodeMirror is available as a JAR file on Maven Central under the
org.webjars group ID.

Add support in the download wrapper to consume CodeMirror from this
location and extract from the JAR file rather than a ZIP file.

The support to download as a ZIP file from the Gerrit bucket is kept
for backward compatibility in case we want to use our own build of
the library at any point in the future.

Change-Id: Ie21b72b57fe1a90894c4aadd7ef4bdc91d055cd5

Former-commit-id: f326c904ff922ff41cee212b409bab2ae5631b7d