Improve performance of ListBranches REST endpoint
Branch listing was reading all the refs from the repo and then
filtering out any branches that were not refs/meta/config or starting
with refs/heads. This extra IO and processing is hard to notice on fast
file systems but it's not on slow ones, especially if repo has a lot of
refs that do not start with refs/heads (e.g. refs/changes).

Rework the branch listing to get only the refs that start with refs/heads
from the repo and then add the 2 other branches required: HEAD and
refs/meta/config.

In one of our repos with 10k branches and 250k changes, this change
reduced the branch listing to 10% of the time it took without this
change.

Change-Id: I0932d06229e6a8cbc9497abcc5b604d2f9a0113b

Former-commit-id: 368d234ac6c6b4e47cb992a4914498731fb9dc3e