Update to Prolog Cafe 1.4
Prolog Cafe is now built with Buck and has a slightly smaller
runtime. Many features not required for Gerrit Code Review were
stripped out of the fork.

Package names were reorganized to make the runtime smaller to
read by pulling exceptions out to their own package. This is
a breaking change for any plugins that contribute a predicate.

Change-Id: Icb50d306dfda146497814c65e2c03fae416ec5e1

Former-commit-id: cc349904933c20509e390b56d3b04b9a6fc51a1d