Fix F5 shortcut being hijacked by edit topic in Firefox
In change screen, if the F5 key was pressed the topic text field
opened instead of refreshing the page. In Google Chrome, F5 works
as intended.

The problem seems to be the way GWT manages the keypress events for
some keys in different browsers and an open issue addresses this:
https://code.google.com/p/google-web-toolkit/issues/detail?id=1061

This patch is a workaround to avoid the problem in Firefox and all
Gecko based browsers until the GWT issue is resolved.

Bug: Issue 2815
Change-Id: I41a4c117d39db7437ed352d10a804d8e8ada8f07

Former-commit-id: ba87b42804f0dfeccc27abb21eee8cfef8f19ec3