Start using TestRepository instead of Git in tests
The add, rm, and checkout methods in GitUtil (among others) require a
non-bare repository, which implies local disk. In order to make tests
faster, we would like to move these repositories entirely into bare
InMemoryRepositorys. As a start, convert PushOneCommit to store a
CommitBuilder and use its add/rm methods instead of depending on the
local filesystem to store state. Instead of checkout, use the new
reset method on TestRepository.

Change-Id: Icb286719bc99dae841f2a234677524abd7f144dc

Former-commit-id: 7dadbcf3f172b0d44c2548c42fdf8fe1ff920977