Fix jar/non-jar plugin ordering on java.nio.Path
When I5212d248 has been merged the plugin loading order has been screwed
up: the problem behind is the different behaviour of .endsWith on Path
compared to String.

With String the .endsWith actually matches the file extension whilst on
Path it checks the last part of the path element and NOT the extension.

The correct path name matching has to be implemented on file name string

Change-Id: I1485f8b71663518152ab15a1910f30cd10a5801f
Former-commit-id: 5ebdd3f208614d6ce50d7e0d5415184c3f3e91c0