Use an annotation for project configs in acceptance tests
AbstractSubmit, among other tests, was double-creating projects (not a
very fast operation) just for the purposes of avoiding empty commits
in some cases. Instead of this ad-hoc, SSH-based approach, specify
project configuration in an annotation like we do with server
configuration. This annotation is directly converted to a
ProjectInput, and we use the in-process API instead of requiring an
SSH connection for this step.

Change-Id: I45a98939cdcbe6a486d9b62a14c52dc53aaa5fd7

Former-commit-id: 4ff91616a1fa2c79cef3b5e78b3f057edcc2c7de