Use distinct project names for each acceptance test method
The project name does not need to be a known value, as it can always
be looked up using the project field. Reusing a project name prevents
us from reusing a single running server for multiple test methods,
which would improve performance.

This requires relatively minor changes to tests, with the in
retrospect obvious exception of ListProjectsIT.

Change-Id: I2c19b444e005a2783d28d8182a65d36208175ac9

Former-commit-id: 4348b80dd8d8024cab81c606dbc675d61b68a772