Allow scrollable cover messages
Previously, if a line started with double spaces
(monospacing/.wikiPreFormat) it could seriously mess up the
comment layout if being too long.

And since code snippets are usually posted this way, we do
not want to wrap the lines, since it could hinder readability.

This change allows the cover message to be horizontally
scrollable in those cases.

Change-Id: I7ae9134ef60eff36fc7684ec31d639eeae8d6570

Former-commit-id: 440650445ee8a8eb9654c230f231d3484a0cf647