OAuth: Allow to change username
OAuth extension point was derived from GitHub OAuth implementation.
Changing of user setting was disabled for GitHub, because it can be
induced from the endpoints. Given that other OAuth providers don't
expose username it makes sense to allow to change the attributes,
as it is the case for OpenID authentication scheme.

Inspired-by: Kelly Campbell <kelly.a.campbell@gmail.com>
Change-Id: Idc15271b619f11c17c9ea88e611259e5a113b0e5

Former-commit-id: 735f7fa45831679ff378ce53c84a53633ebf7ac2