Move common code from ActionsIT and AbstractSubmit to AbstractDaemonTest
Move the submit and getActions methods out of ActionsIT into
AbstractDaemonTest so they can be reused by all the subclassed tests.

Remove the duplicated submit method from AbstractSubmit.

This reverts change Ic60d2aea which made the submit and getActions methods
static in ActionsIT. With that change it was necessary to make ActionsIT
a dependency of any other test class that wanted to reuse the submit
or getActions methods.

Change-Id: I1f25fea22bd75e171bd1619ea7e0b31dba67efb6

Former-commit-id: eecb2c598d9c677a9f574e8b5663b2d88acf736e