ChangeData: check patch set visibility in change mergeability check
If the latest patch set of a change is not visible to the caller (for
example if it's a draft) then the call to getSubmitType fails, causing
an OrmException which then results in an Internal Server Error being
displayed to the caller.

Add a check for patch set visibility before calling getSubmitType.

Bug: Issue 3249
Change-Id: I4ab812dc0c54d81a98b75fceb42fbd203443c549

Former-commit-id: faaa54874a5f764cfd76f5e1888010952244fd13