CherryPick: Append new branch to topic for cherry-picked change
When cherry picking a change to another branch gerrit used to preserve
the topic of the original change. This may lead to confusion if
`change.submitWholeTopic` is enabled as then the newly cherry-picked
change is part of a set of changes which is supposed to be submitted
together.

This seems to be unnecessary as cherry picking usually involves moving
important changes such as bug fixes to another branch which is not
supposed to be part of the same set to be submitted at the same time.

Change-Id: I195774b7f594553018e13101a6b3ce904249eed6

Former-commit-id: c5d7bfd972009307d07c1f6f71ff655126393af7