AbstractDaemonTest: Simplify helpers for getting changes
There were two versions, getChange using HTTP and get using the
extension API. Standardize on the extension API, which works even on
@NoHttpd tests. Update callers to use get() or info() as appropriate.

Remove the version that passes a specific user, as it was unused.
Callers can work around this with setApiUser, or we can always add
back a helper to call setApiUser and unset it after the get call.

Change-Id: Idbbea5f54b4e5098235b2bfd681a2d29e413c81e

Former-commit-id: 429ae9f1a177f594f115f87524944c914a6a71fb