Workaround Guice bug "getPathInfo not decoded"
Due to the Guice issue [1] cloning of a repository with a space in its
name was impossible. If the space wasn't encoded:

  $ git clone 'http://gerrit/My Project'

then such a request failed already in Jetty code.
If the space was encoded:

  $ git clone 'http://gerrit/My+Project'

or:

  $ git clone 'http://gerrit/My%20Project'

then the HttpServletRequest.getPathInfo in the GitOverHttpServlet would
return a non-decoded string i.e. "My%20Project' and the project wouldn't
be found. The HttpServletRequest.getPathInfo is expected to return a
decoded string [2].

Unfortunately, there is no Guice release which contains the bugfix for
the issue [1].

Workaround: decode the projectName in the GitOverHttpServlet.Resolver
explicitly.

[1] https://github.com/google/guice/issues/745
[2] http://docs.oracle.com/javaee/7/api/javax/servlet/http/HttpServletRequest.html#getPathInfo()

Change-Id: I8872cd2d1bab3f8e454d0d7399b35eb850068042

Former-commit-id: 304130b0674adad07e5c9576bd366b5b1021e1b6